Update client side to be able to start modelserver as eclipse product
- Fix product launch config for glsp product
- Adjust build.sh file (do not offer copy server products anymore as they are automatically published to the client)
- Remove ecore-server package and use theia-ecore instead as intended
- Update README
- Keep modelserver product running
