+++ client/README.md
@@ -19,7 +19,7 @@ Install yarn.
 ## Remarks regarding glsp-server start up
 You will need openjdk >= 11 on your path, as the client starts the ecore-glsp-server.
 
-Before you start the browser or electron application you have to run the build script `build.sh` in the root directory (this is also explained in the README file there). This copies the build artifacts into the build folder of `ecore-server` package.
+Before you start the browser or electron application you have to run the build script `build.sh` in the root directory (this is also explained in the README file there). This builds the server products and publishes them into build folder of `theia-ecore` package.
 
 ## Running the browser example
