--- client/ecore-server/.eslintrc.js
@@ -1,15 +1,0 @@
-/** @type {import('eslint').Linter.Config} */
-module.exports = {
-    extends: [
-        '../configs/base.eslintrc.json',
-        '../configs/warnings.eslintrc.json',
-        '../configs/errors.eslintrc.json'
-    ],
-    ignorePatterns: [
-        '**/{node_modules,lib,server}'
-    ],
-    parserOptions: {
-        tsconfigRootDir: __dirname,
-        project: 'tsconfig.json'
-    }
-};
