--- client/ecore-server/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "@eclipse-emfcloud/ecore-server",
-  "keywords": [
-    "theia-extension"
-  ],
-  "author": {
-    "name": "EclipseSource"
-  },
-  "license": "(EPL-2.0 OR MIT)",
-  "version": "0.1.0",
-  "files": [
-    "lib",
-    "src"
-  ],
-  "homepage": "https://github.com/eclipse-emfcloud/ecore-glsp",
-  "bugs": "https://github.com/eclipse-emfcloud/ecore-glsp",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/eclipse-emfcloud/ecore-glsp.git"
-  },
-  "dependencies": {
-    "@eclipse-emfcloud/modelserver-theia": "next",
-    "@eclipse-emfcloud/theia-ecore": "0.2.0",
-    "@theia/core": "latest",
-    "@theia/process": "latest",
-    "@theia/workspace": "latest"
-  },
-  "devDependencies": {
-    "eslint": "^6.8.0",
-    "rimraf": "^2.6.1",
-    "typescript": "~3.9.2"
-  },
-  "scripts": {
-    "prepare": "yarn run clean && yarn run build",
-    "clean": "rimraf lib",
-    "build": "tsc && yarn run lint",
-    "lint": "eslint -c ./.eslintrc.js --ext .ts ./src",
-    "watch": "tsc -w"
-  },
-  "theiaExtensions": [
-    {
-      "backend": "lib/node/backend-extension"
-    }
-  ]
-}
