--- client/ecore-server/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../configs/base.tsconfig.json",
-  "compilerOptions": {
-    "baseUrl": ".",
-    "rootDir": "src",
-    "outDir": "lib",
-    "skipLibCheck": true
-  },
-  "include": [
-    "src"
-  ]
-}
