+++ client/theia-ecore/src/node/ecore-glsp-server-contribution.ts
@@ -8,17 +8,17 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR MIT
  ********************************************************************************/
-import { EcoreLanguage } from "@eclipse-emfcloud/theia-ecore/lib/common/ecore-language";
 import { BaseGLSPServerContribution } from "@eclipse-glsp/theia-integration/lib/node";
 import { IConnection } from "@theia/languages/lib/node";
 import { inject, injectable, optional } from "inversify";
 import * as net from "net";
 import { createSocketConnection } from "vscode-ws-jsonrpc/lib/server";
 
+import { EcoreLanguage } from "../common/ecore-language";
 import { GLSPLaunchOptions } from "./glsp-server-launcher";
 
 @injectable()
-export class EcoreGLServerContribution extends BaseGLSPServerContribution {
+export class EcoreGLSPServerContribution extends BaseGLSPServerContribution {
 
     readonly id = EcoreLanguage.Id;
     readonly name = EcoreLanguage.Name;
