+++ client/theia-ecore/src/node/glsp-server-launcher.ts
@@ -33,8 +33,10 @@ export class GLSPServerLauncher implements BackendApplicationContribution {
     @inject(ILogger) private readonly logger: ILogger;
 
     initialize(): void {
-        if (!this.launchOptions.isRunning && !this.start()) {
-            this.logError("Error during model server startup");
+        if (!this.launchOptions.isRunning) {
+            if (!this.start()) {
+                this.logError("Error during GLSP server startup");
+            }
         } else {
             this.logInfo("GLSP Server is already running");
         }
