+++ server/org.eclipse.emfcloud.ecore.glsp-app/org.eclipse.emfcloud.ecore.glsp.product/org.eclipse.emfcloud.ecore.glsp.product.launch
@@ -1,7 +1,14 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <launchConfiguration type="org.eclipse.pde.ui.RuntimeWorkbench">
-    <setAttribute key="additional_plugins"/>
+    <setAttribute key="additional_plugins">
+        <setEntry value="org.eclipse.core.runtime:3.18.0.v20200506-2143:default:true:default:true"/>
+        <setEntry value="org.eclipse.equinox.common:3.12.0.v20200504-1602:default:true:2:true"/>
+        <setEntry value="org.eclipse.equinox.ds:1.6.200.v20200422-1833:default:true:2:true"/>
+        <setEntry value="org.eclipse.equinox.simpleconfigurator:1.3.500.v20200211-1505:default:true:1:true"/>
+        <setEntry value="org.eclipse.osgi:3.15.300.v20200520-1959:default:true:default:true"/>
+    </setAttribute>
     <booleanAttribute key="append.args" value="true"/>
+    <stringAttribute key="application" value="org.eclipse.emfcloud.ecore.glsp.app.ecore-glsp-server"/>
     <booleanAttribute key="askclear" value="true"/>
     <booleanAttribute key="automaticAdd" value="true"/>
     <booleanAttribute key="automaticValidate" value="true"/>
@@ -18,20 +25,41 @@
     <stringAttribute key="location" value="${workspace_loc}/../runtime-org.eclipse.emfcloud.ecore.glsp.product"/>
     <booleanAttribute key="org.eclipse.jdt.launching.ATTR_USE_START_ON_FIRST_THREAD" value="true"/>
     <stringAttribute key="org.eclipse.jdt.launching.JRE_CONTAINER" value="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-11"/>
-    <stringAttribute key="org.eclipse.jdt.launching.PROGRAM_ARGUMENTS" value="-os ${target.os} -ws ${target.ws} -arch ${target.arch} -nl ${target.nl} -consoleLog"/>
+    <stringAttribute key="org.eclipse.jdt.launching.PROGRAM_ARGUMENTS" value="-os ${target.os} -ws ${target.ws} -arch ${target.arch} -nl ${target.nl} -consoleLog -nosplash"/>
     <stringAttribute key="org.eclipse.jdt.launching.SOURCE_PATH_PROVIDER" value="org.eclipse.pde.ui.workbenchClasspathProvider"/>
     <stringAttribute key="pde.version" value="3.3"/>
     <stringAttribute key="product" value="org.eclipse.emfcloud.ecore.glsp.app.product"/>
+    <stringAttribute key="productFile" value="/org.eclipse.emfcloud.ecore.glsp.product/org.eclipse.emfcloud.ecore.glsp.product"/>
     <setAttribute key="selected_features">
         <setEntry value="org.eclipse.core.runtime.feature:default"/>
-        <setEntry value="org.eclipse.emf.codegen.ecore:default"/>
-        <setEntry value="org.eclipse.emf.codegen:default"/>
-        <setEntry value="org.eclipse.emf.common:default"/>
-        <setEntry value="org.eclipse.emf.ecore:default"/>
-        <setEntry value="org.eclipse.emfcloud.ecore.codegen.feature:default"/>
-        <setEntry value="org.eclipse.emfcloud.ecore.modelserver.feature:default"/>
+        <setEntry value="org.eclipse.emfcloud.ecore.glsp.feature:default"/>
         <setEntry value="org.eclipse.equinox.core.feature:default"/>
-        <setEntry value="org.eclipse.equinox.executable:default"/>
+    </setAttribute>
+    <setAttribute key="selected_target_bundles">
+        <setEntry value="org.apache.felix.gogo.command@default:default"/>
+        <setEntry value="org.apache.felix.gogo.runtime@default:default"/>
+        <setEntry value="org.apache.felix.gogo.shell@default:default"/>
+        <setEntry value="org.apache.felix.scr@default:default"/>
+        <setEntry value="org.eclipse.core.contenttype@default:default"/>
+        <setEntry value="org.eclipse.core.expressions@default:default"/>
+        <setEntry value="org.eclipse.core.jobs@default:default"/>
+        <setEntry value="org.eclipse.core.runtime@default:true"/>
+        <setEntry value="org.eclipse.core.variables@default:default"/>
+        <setEntry value="org.eclipse.equinox.app@default:default"/>
+        <setEntry value="org.eclipse.equinox.common@2:true"/>
+        <setEntry value="org.eclipse.equinox.console@default:default"/>
+        <setEntry value="org.eclipse.equinox.ds@2:true"/>
+        <setEntry value="org.eclipse.equinox.preferences@default:default"/>
+        <setEntry value="org.eclipse.equinox.registry@default:default"/>
+        <setEntry value="org.eclipse.equinox.simpleconfigurator@1:true"/>
+        <setEntry value="org.eclipse.osgi.compatibility.state@default:default"/>
+        <setEntry value="org.eclipse.osgi.services@default:default"/>
+        <setEntry value="org.eclipse.osgi.util@default:default"/>
+        <setEntry value="org.eclipse.osgi@default:true"/>
+    </setAttribute>
+    <setAttribute key="selected_workspace_bundles">
+        <setEntry value="org.eclipse.emfcloud.ecore.glsp.app@default:default"/>
+        <setEntry value="org.eclipse.emfcloud.ecore.glsp@default:default"/>
     </setAttribute>
     <booleanAttribute key="show_selected_only" value="false"/>
     <stringAttribute key="templateConfig" value="${target_home}/configuration/config.ini"/>
